Internal

PowerSimulationNODE.MassMatrix — Method

Makes a Float64 Mass Matrix of ones for the ODEProblem. Takes # of differential and algebraic states

PowerSimulationNODE.Source_to_function_of_time — Method

PVS_to_function_of_time(source::PeriodicVariableSource)

Takes in a PeriodicVariableSource from PowerSystems and generates functions of time for voltage magnitude and angle

PowerSimulationNODE.activate_next_source! — Method

activate_next_source!(sys::System)

Either activate the first source if none are available, or make the next source available. To be used in training surrogate to move on to the next system disturbance. Returns the available source

PowerSimulationNODE.build_sys_init — Method

Test function description